backend/requirements.txt
Django==1.11.7
django-rest-framework==0.1.0
djangorestframework==3.7.3
drf-extensions==0.3.1
Jinja2==2.10
MarkupSafe==1.0
olefile==0.44
Pillow==4.3.0
PyMySQL==0.7.11
pytz==2017.3
